Move account lib to user lib, clean old lib
