--- libs/account/jest.config.js
@@ -1,5 +1,0 @@
-module.exports = {
-  name: 'account',
-  preset: '../../jest.config.js',
-  coverageDirectory: '../../coverage/libs/account'
-};
