--- libs/account/package.json
@@ -1,8 +1,0 @@
-{
-  "name": "@account/account",
-  "version": "0.0.1",
-  "peerDependencies": {
-    "@angular/common": "^6.0.0-rc.0 || ^6.0.0",
-    "@angular/core": "^6.0.0-rc.0 || ^6.0.0"
-  }
-}
