--- libs/account/src/index.ts
@@ -1,6 +1,0 @@
-export * from './lib/account/account.module';
-export * from './lib/profile/profile.module';
-
-// Component Module
-export * from './lib/profile/profile-menu/profile-menu.module';
-export * from './lib/profile/profile-widget/profile-widget.module';
