--- libs/account/src/lib/profile/index.ts
@@ -1,1 +1,0 @@
-export * from './profile-widget/profile-widget.module';
