--- libs/account/src/test-setup.ts
@@ -1,2 +1,0 @@
-import 'jest-preset-angular';
-import 'jest';
