--- libs/account/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "types": ["jest"]
-  },
-  "include": ["**/*.ts"]
-}
