--- libs/account/tslint.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../../tslint.json",
-  "rules": {
-    "directive-selector": [true, "attribute", ["profile", "user", "account"], "camelCase"],
-    "component-selector": [true, "element", ["profile", "user", "account"], "kebab-case"]
-  }
-}
