+++ libs/user/src/lib/user/index.ts
@@ -1,0 +1,6 @@
+export * from './account/account.module';
+export * from './profile/profile.module';
+
+// Component Module
+export * from './profile/profile-menu/profile-menu.module';
+export * from './profile/profile-widget/profile-widget.module';
