+++ libs/user/src/lib/user/profile/index.ts
@@ -1,0 +1,2 @@
+export * from './profile-widget/profile-widget.module';
+export * from './profile-menu/profile-menu.module';
