+++ libs/user/src/lib/user/profile/profile-menu/profile-menu.component.scss
