+++ libs/user/src/lib/user/profile/profile-widget/profile-widget.component.scss
